+++ lib/shoes.rb
@@ -91,8 +91,7 @@ require 'shoes/font'
 require 'shoes/gradient'
 require 'shoes/image'
 require 'shoes/image_pattern'
-require 'shoes/keypress'
-require 'shoes/keyrelease'
+require 'shoes/key_event'
 require 'shoes/line'
 require 'shoes/list_box'
 require 'shoes/logger'
