+++ lib/shoes/swt.rb
@@ -80,8 +80,7 @@ require 'shoes/swt/gradient'
 require 'shoes/swt/image'
 require 'shoes/swt/image_pattern'
 require 'shoes/swt/key_event'
-require 'shoes/swt/keypress'
-require 'shoes/swt/keyrelease'
+require 'shoes/swt/key_listener'
 require 'shoes/swt/line'
 require 'shoes/swt/list_box'
 require 'shoes/swt/oval'
