--- lib/shoes/swt/keyrelease.rb
@@ -1,9 +1,0 @@
-class Shoes
-  module Swt
-    class Keyrelease < Keyevent
-      def initialize dsl, app, &blk
-        super(dsl, app, :released, &blk)
-      end
-    end
-  end
-end
